requirements.txt
opencv-python
numpy
ultralytics
pathfinding
Pillow
customtkinter